v0.2.1 [making manual-menu]

+++ app/src/main/java/com/example/xrhstos/bookapp/Book.java
@@ -245,7 +245,7 @@ public class Book implements Parcelable {
     return personalRating;
   }
 
-  public void setPersonalRating(int personalRating) {
+  public void setPersonalRating(float personalRating) {
     this.personalRating = personalRating;
   }
